fix : moved setupTokens to testData
